move partition refs into it's own crate
