+++ src/daft-distributed/Cargo.toml
@@ -19,6 +19,7 @@ daft-local-plan = {path = "../daft-local-plan", default-features = false}
 daft-logical-plan = {path = "../daft-logical-plan", default-features = false}
 daft-scan = {path = "../daft-scan", default-features = false}
 daft-micropartition = {path = "../daft-micropartition", default-features = false}
+daft-partition-refs = {path = "../daft-partition-refs", default-features = false}
 daft-recordbatch = {path = "../daft-recordbatch", default-features = false}
 daft-schema = {path = "../daft-schema", default-features = false}
 futures = {workspace = true}
@@ -44,6 +45,7 @@ python = [
   "common-treenode/python",
   "daft-logical-plan/python",
   "daft-local-plan/python",
+  "daft-partition-refs/python",
   "daft-scan/python",
   "daft-context/python"
 ]
