+++ src/daft-distributed/src/python/ray/mod.rs
@@ -3,8 +3,8 @@ mod worker;
 mod worker_manager;
 
 use common_error::DaftResult;
+pub use daft_partition_refs::RayPartitionRef;
 use pyo3::prelude::*;
-pub use task::RayPartitionRef;
 pub(crate) use task::{RaySwordfishTask, RayTaskResult};
 pub(crate) use worker::RaySwordfishWorker;
 pub(crate) use worker_manager::RayWorkerManager;
