+++ src/daft-local-execution/Cargo.toml
@@ -20,7 +20,6 @@ daft-core = {path = "../daft-core", default-features = false}
 daft-groupby = {path = "../daft-groupby", default-features = false}
 daft-context = {path = "../daft-context", default-features = false}
 daft-csv = {path = "../daft-csv", default-features = false}
-daft-distributed = {path = "../daft-distributed", default-features = false}
 daft-dsl = {path = "../daft-dsl", default-features = false}
 daft-functions-list = {path = "../daft-functions-list", default-features = false}
 daft-functions-uri = {workspace = true}
@@ -31,6 +30,7 @@ daft-local-plan = {path = "../daft-local-plan", default-features = false}
 daft-logical-plan = {path = "../daft-logical-plan", default-features = false}
 daft-micropartition = {path = "../daft-micropartition", default-features = false}
 daft-parquet = {path = "../daft-parquet", default-features = false}
+daft-partition-refs = {path = "../daft-partition-refs", default-features = false}
 daft-recordbatch = {path = "../daft-recordbatch", default-features = false}
 daft-scan = {path = "../daft-scan", default-features = false}
 daft-shuffles = {path = "../daft-shuffles", default-features = false}
@@ -77,6 +77,7 @@ python = [
   "daft-local-plan/python",
   "daft-logical-plan/python",
   "daft-micropartition/python",
+  "daft-partition-refs/python",
   "daft-scan/python",
   "daft-writers/python",
   "daft-text/python"
