+++ src/daft-local-plan/Cargo.toml
@@ -2,7 +2,6 @@
 common-metrics = {path = "../common/metrics", default-features = false}
 common-error = {path = "../common/error", default-features = false}
 common-io-config = {path = "../common/io-config", default-features = false}
-common-partitioning = {path = "../common/partitioning", default-features = false}
 common-py-serde = {path = "../common/py-serde", default-features = false}
 common-resource-request = {path = "../common/resource-request", default-features = false}
 common-treenode = {path = "../common/treenode", default-features = false}
@@ -10,6 +9,7 @@ daft-core = {path = "../daft-core", default-features = false}
 daft-dsl = {path = "../daft-dsl", default-features = false}
 daft-logical-plan = {path = "../daft-logical-plan", default-features = false}
 daft-micropartition = {path = "../daft-micropartition", default-features = false}
+daft-partition-refs = {path = "../daft-partition-refs", default-features = false}
 daft-scan = {path = "../daft-scan", default-features = false}
 daft-recordbatch = {path = "../daft-recordbatch", default-features = false}
 bincode = {workspace = true}
@@ -34,6 +34,7 @@ python = [
   "daft-dsl/python",
   "daft-logical-plan/python",
   "daft-micropartition/python",
+  "daft-partition-refs/python",
   "daft-scan/python"
 ]
