+++ src/daft-partition-refs/Cargo.toml
@@ -1,0 +1,20 @@
+[dependencies]
+common-partitioning = {path = "../common/partitioning", default-features = false}
+common-py-serde = {path = "../common/py-serde", default-features = false, optional = true}
+pyo3 = {workspace = true, optional = true}
+serde = {workspace = true}
+
+[features]
+python = [
+  "dep:common-py-serde",
+  "dep:pyo3",
+  "common-partitioning/python"
+]
+
+[lints]
+workspace = true
+
+[package]
+edition = {workspace = true}
+name = "daft-partition-refs"
+version = {workspace = true}
